+++ www/views/Extjs3/javascript/MainLayout.js
@@ -470,7 +470,9 @@ Ext.extend(GO.MainLayout, Ext.util.Observable, {
 			}),
 			listeners: {
 				show: function() {
-					this.startMenuSearchField.focus(false, 500);
+					if(!GO.util.isMobileOrTablet()) {
+						this.startMenuSearchField.focus(false, 500);
+					}
 				},
 				itemclick : function(item, e) {
 					if (!item.textField) {