+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - WOPI: Two attachments with same name caused problems when opening via WOPI (Libreoffice or o365)
 - OAuth2Client / Email: Refactor renewing of access token, better check access token on sending email
 - ActiveSync: All day events could shift when made with phone in another timezone
+- Core: turning mobile didn't resize screen correctly
 
 17-11-2022 6.6.131
 - Wrong build