+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Core: background color when button is focussed
 - WOPI: Two attachments with same name caused problems when opening via WOPI (Libreoffice or o365)
 - OAuth2Client / Email: Refactor renewing of access token, better check access token on sending email
+- OAuth2Client: Better check for expired Refresh token
 
 17-11-2022 6.6.131
 - Wrong build