+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Core: turning mobile didn't resize screen correctly
 - Core: don't focus on search in start menu as it causes it to close on android phones
 - Sync: Allow deleting of items by default again
+- Core: install without ioncube
 
 17-11-2022 6.6.131
 - Wrong build